clean up error types, moving CompactFormatError to error module and replacing ZCB error type with new internal EncodingInvalid error type
